agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn",
    "pydantic-ai-slim[ag-ui]",
    "pydantic-ai-slim[openai]",
    "python-dotenv",
    "logfire>=4.10.0",
    "zep-cloud",
]
